Improving SimpleITK resample function.
Before we had two fuction one to load the segmentation and one to load the
MRI sequence and resample to a new pixel spacing. These functions were not working for all ADC and HBV co-registration images, tested on PI-CAI.

Fix: the two fuction have been merged into one function that gets a boolean depending on the type of image (mask or not). The updated version also
resolves some issues regarding resampled.